requirements_dev.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
Werkzeug==3.0.3
# psycopg2-binary removed for dev - use SQLite
# Add back for production/Neon

